Extract core automation models
