+++ docs/DECISIONS.md
@@ -10,6 +10,7 @@
 - TrustedRouter model defaults and catalog normalization live in focused core files, not the general `Models.swift` domain bucket. `TrustedRouterDefaults.swift` owns provider IDs, aliases, bundled fallback rows, Nike 1.0/Prometheus 1.0 branding, and sort/category policy; `ModelInfo.swift` owns the catalog value records.
 - App configuration is a focused core boundary. `AppConfig.swift` owns persisted model/mode/base-URL settings, OAuth/developer-override compatibility, signed-in TrustedRouter account metadata, and favorite-model normalization; `Models.swift` stays focused on general thread/project records.
 - Core tool records are a focused schema boundary. `ToolModels.swift` owns `ToolDefinition`, `ToolCall`, `ToolResult`, built-in core tool definitions, tool-call transcript redaction, and tool-specific browser/memory output compatibility; general chat/thread/project models should not own tool payload compatibility rules.
+- Core automation records are a focused scheduling boundary. `AutomationModels.swift` owns automation kind/status/schedule enums, recurrence interval semantics, next-run calculation, and display sorting; general chat/thread/project models should not own recurring-workflow scheduling rules.
 - Tool-router feature families should delegate before they sprawl. `ToolRouter` owns the common shell/file/patch entry point and composes git definitions from `GitToolCallDispatcher`; local git, GitHub PR, and worktree tool-call argument mapping lives in that dispatcher. This keeps the shared router small while preserving one public `ToolRouter.execute` API for the agent runtime.
 - Shell request policy is a tool-family concern. `ShellToolCallDispatcher` owns `host.shell.run` definitions, cwd containment, timeout bounds, environment override validation, and `ShellExecutionRequest` construction; `ToolRouter` only delegates shell calls before handling file and patch primitives.
